+++ apps/enmeshed/lib/core/modals/create_recovery_kit/save_or_print_recovery_kit.dart
@@ -10,186 +10,23 @@ import 'package:go_router/go_router.dart';
 import 'package:identity_recovery_kit/identity_recovery_kit.dart';
 import 'package:pdf/pdf.dart';
 import 'package:printing/printing.dart';
-import 'package:wolt_modal_sheet/wolt_modal_sheet.dart';
 
-import '/core/utils/extensions.dart';
-import '../constants.dart';
-import '../utils/snackbars.dart';
-import '../widgets/widgets.dart';
+import '/core/utils/utils.dart';
+import '../../constants.dart';
+import '../../widgets/widgets.dart';
 
-Future<void> showCreateRecoveryKitModal({required BuildContext context, required String accountId}) async {
-  final pageIndexNotifier = ValueNotifier<int>(0);
-  final passwordNotifier = ValueNotifier<String?>(null);
-
-  await WoltModalSheet.show<void>(
-    useSafeArea: false,
-    context: context,
-    pageIndexNotifier: pageIndexNotifier,
-    onModalDismissedWithDrag: () => context.pop(),
-    onModalDismissedWithBarrierTap: () => context.pop(),
-    showDragHandle: false,
-    pageListBuilder: (context) => [
-      WoltModalSheetPage(
-        trailingNavBarWidget: Padding(
-          padding: const EdgeInsets.only(right: 8),
-          child: IconButton(icon: const Icon(Icons.close), onPressed: () => context.pop()),
-        ),
-        leadingNavBarWidget: Padding(
-          padding: const EdgeInsets.only(left: 24, top: 20),
-          child: Text(context.l10n.identityRecovery_passwordTitle, style: Theme.of(context).textTheme.titleLarge),
-        ),
-        child: _CreatePassword(
-          onContinue: (String enteredPassword) {
-            passwordNotifier.value = enteredPassword;
-            pageIndexNotifier.value++;
-          },
-        ),
-      ),
-      WoltModalSheetPage(
-        hasTopBarLayer: false,
-        child: ValueListenableBuilder<String?>(
-          valueListenable: passwordNotifier,
-          builder: (context, password, _) {
-            return _SaveRecoveryKit(
-              accountId: accountId,
-              password: password,
-              onBackPressed: () => pageIndexNotifier.value = 0,
-            );
-          },
-        ),
-      ),
-    ],
-  );
-}
-
-class _CreatePassword extends StatefulWidget {
-  final void Function(String password) onContinue;
-
-  const _CreatePassword({required this.onContinue});
-
-  @override
-  State<_CreatePassword> createState() => _CreatePasswordState();
-}
-
-class _CreatePasswordState extends State<_CreatePassword> {
-  final form = GlobalKey<FormState>();
-  final TextEditingController _passwordController = TextEditingController();
-  final TextEditingController _confirmPasswordController = TextEditingController();
-  bool _passwordVisible = true;
-  bool _confirmPasswordVisible = true;
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(left: 24, right: 24, bottom: max(MediaQuery.paddingOf(context).bottom, 16)),
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Text(context.l10n.identityRecovery_passwordDescription),
-          Gaps.h24,
-          InformationContainer(title: context.l10n.identityRecovery_passwordAttention),
-          Gaps.h36,
-          Form(
-            key: form,
-            child: Column(
-              children: [
-                _PasswordTextField(
-                  inputControl: _passwordController,
-                  label: context.l10n.identityRecovery_password,
-                  obscureText: _passwordVisible,
-                  onIconTap: () => setState(() => _passwordVisible = !_passwordVisible),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) return context.l10n.identityRecovery_passwordEmptyError;
-                    return null;
-                  },
-                ),
-                Gaps.h24,
-                _PasswordTextField(
-                  inputControl: _confirmPasswordController,
-                  label: context.l10n.identityRecovery_passwordConfirm,
-                  obscureText: _confirmPasswordVisible,
-                  onIconTap: () => setState(() => _confirmPasswordVisible = !_confirmPasswordVisible),
-                  validator: (value) {
-                    if (value == null || value.isEmpty) return context.l10n.identityRecovery_passwordEmptyError;
-                    if (value != _passwordController.text) return context.l10n.identityRecovery_passwordError;
-                    return null;
-                  },
-                ),
-                Gaps.h48,
-                Align(
-                  alignment: Alignment.centerRight,
-                  child: FilledButton(
-                    style: OutlinedButton.styleFrom(minimumSize: const Size(100, 36)),
-                    onPressed: () {
-                      if (form.currentState!.validate()) widget.onContinue(_passwordController.text);
-                    },
-                    child: Text(context.l10n.identityRecovery_startNow),
-                  ),
-                ),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class _PasswordTextField extends StatelessWidget {
-  final TextEditingController inputControl;
-  final String label;
-  final bool obscureText;
-  final VoidCallback onIconTap;
-  final String? Function(String?) validator;
-
-  const _PasswordTextField({
-    required this.inputControl,
-    required this.label,
-    required this.obscureText,
-    required this.onIconTap,
-    required this.validator,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return TextFormField(
-      controller: inputControl,
-      obscureText: obscureText,
-      decoration: InputDecoration(
-        labelText: label,
-        suffixIcon: IconButton(
-          icon: Icon(
-            obscureText ? Icons.visibility : Icons.visibility_off,
-            color: Theme.of(context).primaryColorDark,
-          ),
-          onPressed: onIconTap,
-        ),
-        border: OutlineInputBorder(
-          borderRadius: const BorderRadius.all(Radius.circular(8)),
-          borderSide: BorderSide(color: Theme.of(context).colorScheme.outline),
-        ),
-        focusedBorder: OutlineInputBorder(
-          borderRadius: const BorderRadius.all(Radius.circular(8)),
-          borderSide: BorderSide(color: Theme.of(context).colorScheme.primary),
-        ),
-      ),
-      validator: validator,
-    );
-  }
-}
-
-class _SaveRecoveryKit extends StatefulWidget {
+class SaveOrPrintRecoveryKit extends StatefulWidget {
   final String accountId;
   final String? password;
   final VoidCallback onBackPressed;
 
-  const _SaveRecoveryKit({required this.accountId, required this.password, required this.onBackPressed});
+  const SaveOrPrintRecoveryKit({required this.accountId, required this.password, required this.onBackPressed, super.key});
 
   @override
-  State<_SaveRecoveryKit> createState() => _SaveRecoveryKitState();
+  State<SaveOrPrintRecoveryKit> createState() => _SaveOrPrintRecoveryKitState();
 }
 
-class _SaveRecoveryKitState extends State<_SaveRecoveryKit> {
+class _SaveOrPrintRecoveryKitState extends State<SaveOrPrintRecoveryKit> {
   Uint8List? _pdf;
 
   @override
@@ -247,7 +84,6 @@ class _SaveRecoveryKitState extends State<_SaveRecoveryKit> {
     );
   }
 
-  // Download is not working on iOS simulator (but works on real device)
   Future<void> _downloadFile() async {
     final result = await FilePicker.platform.saveFile(
       fileName: 'identity_recovery_kit.pdf',
@@ -300,12 +136,12 @@ class _SaveRecoveryKitState extends State<_SaveRecoveryKit> {
     if (!mounted) return;
 
     final generatedPdf = await PdfGenerator(
-      headerTitleHexColor: '#CD5038',
-      backgroundHexColor: '#D0E6F3',
-      defaultTextHexColor: '#004F79',
-      borderHexColor: '#006685',
-      labelHexColor: '#006685',
-      addressHexColor: '#0077B6',
+      headerTitleHexColor: '#1A80D9',
+      backgroundHexColor: '#D7E3F8',
+      defaultTextHexColor: '#17428D',
+      borderHexColor: '#17428D',
+      labelHexColor: '#17428D',
+      addressHexColor: '#1A80D9',
       pdfTexts: (
         headerTitle: context.l10n.identityRecovery_pdfHeaderTitle,
         keepSafeText: context.l10n.identityRecovery_pdfKeepSafeText,
