+++ apps/enmeshed/lib/core/modals/modals.dart
@@ -1,5 +1,5 @@
 export 'create_attribute.dart';
-export 'create_recovery_kit.dart';
+export 'create_recovery_kit/create_recovery_kit.dart';
 export 'delete_attribute.dart';
 export 'enter_password/enter_password.dart';
 export 'succeed_attribute.dart';
